[FIX] product: move view of product packaging to product module
When you activate the use of products packaging in 'stock' module, it
doesn't change anything in the product view if the app 'purchase' is not
installed.

We have this problem because the insert of the product_packaing view in
the product_template form is done in 'purchase' module instead of
'product' module.

To fix this issue, we had to move the page 'Inventory' from 'stock'
moduile to 'product' module. And the group 'packaging' form 'purchase to
'product' module.