[FIX] account: exchange rate of `price_subtotal_signed`
The field `price_subtotal_signed` contains the price converted in the
company currency. However, we should make sure to use the exchange rate
of the invoice date for its computation.

An use problematic use case is the creation of a deferred revenue from
an invoice (commit d126d521ba3).

opw-702352