[IMP] stock_account: Counter-Part Account
In commit 91673fc4277, the counter-part account is made selectable and
required when changing the price of a product.

However, the same wizard is reused for the "Compute from BOM" feature,
in module `product_extended`. This forces the user to set an account
even if the product valuation is not "Perpetual (automated)".

In this commit, the field becomes required only if the product valuation
is "Perpetual (automated)".

opw-697866