[FIX] account_asset: To ensure the migration compatibility with 9.0
This fix allow to keep the same behavior of 8.0 and 10.0 to manage the assets.

In 9.0, the field account_income_recognition_id(oldname='account_expense_depreciation_id') has been
deleted from the account.asset.category.form view with e46f1a2c6ef6b2649c57ef5435dfee604bc7cd84 and
then it was not possible to manage the assets like in 8.0 and 10.0

But as this fix is in a stable version, it must be compatible with instances without any upgrade
of the view account.asset.category.form.

This issue has been solved in saas-11 with this commit: ad2f99bbc9e9112c88cd6bd2d83c5c1712fdc5fe

opw:703860,702265