[FIX] bus: do not suggest company users to portal users
When you are connected as a portal user, you can access to the list of
company users when you try to create a channel of type "Direct Message".

This is caused by a SQL request that doesn't check the access rights of
the user before returning the list of users.

To fix this, we check that the user can create a mail channel, before
returning partners information.

opw-704127