[FIX] point_of_sale: avoid scrollbar on main window
Scrollbar appeared in Firefox (at least 50.1.0 and up). Related to
depending on non-standard behavior when setting height on a table. For a
detailed explanation see 64423337b83570855d019bcea9e9f1a4acb316e9.

In this case we can simply ignore overflow on the main POS window since
the height difference introduced is only a couple of pixels.